chore: use uint256 for index

+++ src/contracts/multichain/ECDSACertificateVerifierStorage.sol
@@ -40,7 +40,7 @@ abstract contract ECDSACertificateVerifierStorage is IECDSACertificateVerifier {
     mapping(bytes32 operatorSetKey => mapping(uint32 referenceTimestamp => uint256 numOperators)) internal _numOperators;
 
     /// @dev Mapping from operatorSetKey to referenceTimestamp to operatorInfos
-    mapping(bytes32 operatorSetKey => mapping(uint32 referenceTimestamp => mapping(uint32 => ECDSAOperatorInfo)))
+    mapping(bytes32 operatorSetKey => mapping(uint32 referenceTimestamp => mapping(uint256 => ECDSAOperatorInfo)))
         internal _operatorInfos;
 
     // Construction
